+++ docs/sources/whatsnew.rst
@@ -9,7 +9,7 @@ Version 1.5.0
 
 The two big changes are:
 
-1. the ability to use `Yaml <yaml-list.html`__ files to specify samples,
+1. the ability to use `Yaml <yaml-list.html>`__ files to specify samples,
 2. the introduction of ``run_for_all`` (and ``run_for_all_samples``) functions to simplify the usage of the ``parallel`` module (see `standard library docs <stdlib.html>`__).
 
 Several of the other changes were then to support these two features.